requirements.txt
wheel==0.42.0
setuptools==65.5.1
numpy==1.23.0
scikit-learn==1.3.0
zope.interface==6.1
pandas==1.5.3
pyarrow==10.0.1
xarray==2023.2.0
shapely==2.0.1
zope.event==5.0
tensorflow==2.11.0
keras==2.11.0
xgboost==1.7.3
matplotlib==3.7.1
windrose==1.9.0
geopandas==0.12.2
netCDF4==1.6.0
feature-engine==1.5.2